refactor: TodoItem 컴포넌트의 렌더링 로직 분리로 가독성 향상
