+++ src/components/Todo/TodoItem.tsx
@@ -1,7 +1,6 @@
 import { Check, Close, Delete, Edit } from '@/assets/icons';
 import { Button } from '@/components/common';
 import IconWrapper from '@/components/IconWrapper';
-import { checkedStyle, checkIconStyle, contentWrapperStyle, todoItemStyle } from '@/components/Todo/TodoList';
 import { TODO_STATUS } from '@/constants/todo';
 import { useToastHandler } from '@/hooks';
 import { useDeleteTodo, useUpdateTodo } from '@/hooks/mutations';
@@ -98,78 +97,139 @@ const TodoItem = ({ todo, squadMemberId }: { todo: ToDoDetail; squadMemberId: nu
     setIsDeleteMode(false);
   };
 
-  return (
-    <>
-      {!isDeleteMode ? (
-        <li css={[todoItemStyle, getStatusStyles(isCompleted, theme)]}>
-          <button onClick={toggleTodoStatus} style={fullSizeButtonStyle} aria-label="투두 상태 변경">
-            <div css={contentWrapperStyle}>
-              <IconWrapper
-                aria-label={isCompleted ? '완료된 투두' : '완료되지 않은 투두'}
-                aria-checked={isCompleted}
-                role="checkbox"
-                css={[checkIconStyle, isCompleted && checkedStyle]}
-              >
-                {isCompleted && <Check />}
-              </IconWrapper>
-              {!isEditMode ? (
-                <p css={isCompleted && contentStyle(theme)}>{contents}</p>
-              ) : (
-                <input
-                  type="text"
-                  value={newContents}
-                  onChange={(e) => setNewContents(e.target.value)}
-                  onClick={(e) => e.stopPropagation()}
-                  onKeyDown={handleKeyPressForEdit}
-                  css={editInputStyle}
-                  autoFocus
-                />
-              )}
-            </div>
-          </button>
-          {!isEditMode ? (
-            <div css={actionStyle} role="presentation">
-              <IconWrapper onClick={() => setIsEditMode(true)} aria-label="투두 수정">
-                <Edit />
-              </IconWrapper>
-              <IconWrapper onClick={handleDeleteTodo} aria-label="투두 삭제">
-                <Delete />
-              </IconWrapper>
-            </div>
-          ) : (
-            <div css={editActionStyle} role="presentation">
-              <Button id="edit-btn" text="수정" variant="confirm" onClick={handleEditContents} aria-label="수정하기" />
-              <Button
-                text="취소"
-                variant="default"
-                onClick={() => {
-                  setNewContents(contents);
-                  setIsEditMode(false);
-                }}
-                aria-label="투두 수정 취소"
-              />
-            </div>
-          )}
-        </li>
+  if (isDeleteMode) {
+    return (
+      <li css={[todoItemStyle, deleteModeStyle]}>
+        <Button onClick={handleConfirmDelete} text="삭제" css={deleteButtonStyle} aria-label="삭제" />
+        <IconWrapper css={cancelButtonStyle} onClick={() => setIsDeleteMode(false)} aria-label="삭제 취소">
+          <Close />
+        </IconWrapper>
+      </li>
+    );
+  }
+
+  const renderTodoContent = () => (
+    <div css={contentWrapperStyle}>
+      <IconWrapper
+        aria-label={isCompleted ? '완료된 투두' : '완료되지 않은 투두'}
+        aria-checked={isCompleted}
+        role="checkbox"
+        css={[checkIconStyle, isCompleted && checkedStyle]}
+      >
+        {isCompleted && <Check />}
+      </IconWrapper>
+      {isEditMode ? (
+        <input
+          type="text"
+          value={newContents}
+          onChange={(e) => setNewContents(e.target.value)}
+          onClick={(e) => e.stopPropagation()}
+          onKeyDown={handleKeyPressForEdit}
+          css={editInputStyle}
+          autoFocus
+        />
       ) : (
-        <li css={[todoItemStyle, deleteModeStyle]}>
-          <Button
-            onClick={handleConfirmDelete}
-            text="삭제"
-            style={fullSizeButtonStyle}
-            css={deleteButtonStyle}
-            aria-label="삭제"
-          />
-          <IconWrapper css={cancelButtonStyle} onClick={() => setIsDeleteMode(false)} aria-label="삭제 취소">
-            <Close />
-          </IconWrapper>
-        </li>
+        <p css={isCompleted && completedContentStyle} style={{ textAlign: 'left' }}>
+          {contents}
+        </p>
       )}
-    </>
+    </div>
+  );
+
+  const renderDefaultActions = () => (
+    <div css={actionStyle} role="presentation">
+      <IconWrapper onClick={() => setIsEditMode(true)} aria-label="투두 수정">
+        <Edit />
+      </IconWrapper>
+      <IconWrapper onClick={handleDeleteTodo} aria-label="투두 삭제">
+        <Delete />
+      </IconWrapper>
+    </div>
+  );
+
+  const renderEditActions = () => (
+    <div css={editActionStyle} role="presentation">
+      <Button id="edit-btn" text="수정" variant="confirm" onClick={handleEditContents} aria-label="수정하기" />
+      <Button
+        text="취소"
+        variant="default"
+        onClick={() => {
+          setNewContents(contents);
+          setIsEditMode(false);
+        }}
+        aria-label="투두 수정 취소"
+      />
+    </div>
+  );
+
+  return (
+    <li css={[todoItemStyle, getStatusStyles(isCompleted, theme)]}>
+      <button onClick={toggleTodoStatus} css={fullSizeButtonStyle} aria-label="투두 상태 변경">
+        {renderTodoContent()}
+      </button>
+      {isEditMode ? renderEditActions() : renderDefaultActions()}
+    </li>
   );
 };
 
 export default TodoItem;
+
+export const todoItemStyle = css`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 8px;
+  border-radius: 16px;
+  margin: 0 8px;
+  padding: 8px 0 8px 8px;
+  min-height: 48px;
+  flex-shrink: 0;
+  cursor: pointer;
+  transition: transform 0.3s ease-out;
+
+  & button,
+  div {
+    font-size: 0.8rem;
+  }
+
+  ${pcMediaQuery(css`
+    & button,
+    div {
+      font-size: 1rem;
+    }
+  `)}
+`;
+
+export const contentWrapperStyle = css`
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  height: 100%;
+`;
+
+export const checkIconStyle = (theme: Theme) => css`
+  min-width: 20px;
+  max-width: 20px;
+  height: 20px;
+  background-color: ${theme.colors.background.white};
+  border: 2px solid ${theme.colors.gray.gray100};
+  border-radius: 8px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+export const checkedStyle = (theme: Theme) => css`
+  background-color: ${theme.colors.primary};
+  border: none;
+
+  & svg {
+    color: white;
+    stroke-width: 2;
+    width: 12px;
+    height: 12px;
+  }
+`;
 
 export const getStatusStyles = (isChecked: boolean, theme: Theme) => {
   switch (isChecked) {
@@ -185,7 +245,7 @@ export const getStatusStyles = (isChecked: boolean, theme: Theme) => {
   }
 };
 
-const contentStyle = (theme: Theme) => css`
+const completedContentStyle = (theme: Theme) => css`
   color: ${theme.colors.gray.gray200};
   text-decoration: line-through ${theme.colors.gray.gray200};
   text-decoration-thickness: 1.5px;
